refactor: auth와 login 폴더 통합으로 cross feature 의존성 문제 해결
