+++ lib/features/auth/application/use_cases/sign_in_with_google_use_case.dart
@@ -1,5 +1,5 @@
-import 'package:ridingmate/features/login/domain/entities/user.dart';
-import 'package:ridingmate/features/login/domain/repositories/google_auth_repository.dart';
+import 'package:ridingmate/features/auth/domain/entities/user.dart';
+import 'package:ridingmate/features/auth/domain/repositories/google_auth_repository.dart';
 
 class SignInWithGoogleUseCase {
   const SignInWithGoogleUseCase({required GoogleAuthRepository repository})
