+++ lib/features/auth/application/use_cases/sign_in_with_kakao_use_case.dart
@@ -1,5 +1,5 @@
-import 'package:ridingmate/features/login/domain/entities/user.dart';
-import 'package:ridingmate/features/login/domain/repositories/kakao_auth_repository.dart';
+import 'package:ridingmate/features/auth/domain/entities/user.dart';
+import 'package:ridingmate/features/auth/domain/repositories/kakao_auth_repository.dart';
 
 class SignInWithKakaoUseCase {
   const SignInWithKakaoUseCase({required KakaoAuthRepository repository})
