+++ lib/features/auth/application/use_cases/sign_out_with_apple_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:ridingmate/features/login/domain/repositories/apple_auth_repository.dart';
+import 'package:ridingmate/features/auth/domain/repositories/apple_auth_repository.dart';
 
 class SignOutWithAppleUseCase {
   const SignOutWithAppleUseCase({required AppleAuthRepository repository})
