+++ lib/features/auth/application/use_cases/sign_out_with_google_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:ridingmate/features/login/domain/repositories/google_auth_repository.dart';
+import 'package:ridingmate/features/auth/domain/repositories/google_auth_repository.dart';
 
 class SignOutWithGoogleUseCase {
   const SignOutWithGoogleUseCase({required GoogleAuthRepository repository})
