+++ lib/features/auth/data/repositories/apple_auth_repository_impl.dart
@@ -1,7 +1,7 @@
-import 'package:ridingmate/features/login/data/datasources/apple_auth_datasource.dart';
-import 'package:ridingmate/features/login/domain/entities/user.dart';
-import 'package:ridingmate/features/login/domain/enums/login_provider.dart';
-import 'package:ridingmate/features/login/domain/repositories/apple_auth_repository.dart';
+import 'package:ridingmate/features/auth/data/datasources/apple_auth_datasource.dart';
+import 'package:ridingmate/features/auth/domain/entities/user.dart';
+import 'package:ridingmate/features/auth/domain/enums/login_provider.dart';
+import 'package:ridingmate/features/auth/domain/repositories/apple_auth_repository.dart';
 import 'package:sign_in_with_apple/sign_in_with_apple.dart';
 
 class AppleAuthRepositoryImpl implements AppleAuthRepository {
