+++ lib/features/auth/domain/entities/user.dart
@@ -1,4 +1,4 @@
-import 'package:ridingmate/features/login/domain/enums/login_provider.dart';
+import 'package:ridingmate/features/auth/domain/enums/login_provider.dart';
 
 class User {
   const User({
