+++ lib/features/auth/domain/repositories/google_auth_repository.dart
@@ -1,4 +1,4 @@
-import 'package:ridingmate/features/login/domain/entities/user.dart';
+import 'package:ridingmate/features/auth/domain/entities/user.dart';
 
 abstract class GoogleAuthRepository {
   Future<User?> signIn();
